tg_bot/requirements.txt
aiogram
aiohttp
python-dotenv
psycopg2-binary
celery
arq
redis
asyncpg
